Deleting executions
What's done:
* added functionality to delete executions
Closes #383
