+++ save-backend/src/main/kotlin/org/cqfn/save/backend/controllers/ExecutionController.kt
@@ -146,6 +146,7 @@ class ExecutionController(private val executionService: ExecutionService,
         try {
             projectService.findByNameAndOwner(name, owner)?.id?.let {
                 testExecutionService.deleteTestExecutionWithProjectId(it)
+                agentStatusService.deleteAgentStatusWithProjectId(it)
                 agentService.deleteAgentWithProjectId(it)
                 executionService.deleteExecutionByProjectNameAndProjectOwner(name, owner)
             }
